rename and add temp hail methods
